Move internal types to Internal folder.
