+++ src/Faithlife.Ananke/AnankeRunner.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Faithlife.Ananke.Logging;
+using Faithlife.Ananke.Logging.Internal;
 using Faithlife.Ananke.Services;
 using Microsoft.Extensions.Logging;
