+++ src/Faithlife.Ananke/AnankeSettings.cs
@@ -3,8 +3,10 @@ using System.IO;
 using System.Runtime.InteropServices;
 using System.Threading;
 using Faithlife.Ananke.Logging;
+using Faithlife.Ananke.Logging.Internal;
 using Faithlife.Ananke.Services;
 using Microsoft.Extensions.Logging;
+using AnankeLoggerProvider = Faithlife.Ananke.Logging.Internal.AnankeLoggerProvider;
 
 namespace Faithlife.Ananke
 {
