+++ src/Faithlife.Ananke/Logging/Escaping.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using Faithlife.Ananke.Logging.Internal;
 
 namespace Faithlife.Ananke.Logging
 {
