+++ src/Faithlife.Ananke/Logging/Internal/EscapingStringLog.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Faithlife.Ananke.Services;
-
-namespace Faithlife.Ananke.Logging
+﻿namespace Faithlife.Ananke.Logging.Internal
 {
 	/// <summary>
 	/// A string logger that backslash-escapes EOL characters before passing them to an inner logger. This type is threadsafe if its underlying <see cref="IStringLog"/> is threadsafe.
